Remove loginPopup function from app component. Added loginPopup to home page. Got login pop up to show up on server
