+++ src/pages/home/home.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
-import { NavController } from 'ionic-angular';
+import { NavController, AlertController } from 'ionic-angular';
 import { MapPage} from '../map/map';
+import { Angular2TokenService } from 'angular2-token';
 
 @Component({
   selector: 'page-home',
@@ -8,12 +9,53 @@ import { MapPage} from '../map/map';
 })
 export class HomePage {
 
-  constructor(public navCtrl: NavController) {
+  currentUser: any;
 
+  constructor(public navCtrl: NavController, private _tokenService: Angular2TokenService,public alertCtrl: AlertController,) {
+    this._tokenService.init({
+      apiBase: 'https://ufoods.herokuapp.com/'
+    });
   }
 
   goToMapPage() {
     this.navCtrl.push(MapPage);
   }
-
+  loginPopUp() {
+    let confirm = this.alertCtrl.create({
+      title: 'Login',
+      inputs: [
+        {
+          name: 'email',
+          placeholder: 'email'
+        },
+        {
+          name: 'password',
+          placeholder: 'password',
+          type: 'password'
+        }
+      ],
+      buttons: [
+        {
+          text: 'Cancel',
+          handler: data => {
+          }
+        },
+        {
+          text: 'Login',
+          handler: data => {
+            this.login(data);
+          }
+        }
+      ]
+    });
+    confirm.present();
+  }
+  login(credentials) {
+    this._tokenService
+    .signIn(credentials)
+    .subscribe(
+      res => (this.currentUser = res.json().data),
+      err => console.error('error')
+    );
+  }
 }
