+++ src/providers/user/user.ts
@@ -1,5 +1,4 @@
 import { Injectable } from '@angular/core';
-import { Http } from '@angular/http';
 import 'rxjs/add/operator/map';
 import { Angular2TokenService } from 'angular2-token';
